refactor: use clap as a commands parser (#3867)
Description
---
The PR refactors replace our own parser with a derived by the `clap` crate.
It also adds the `watch` command, and avoids unexpected status printing. 

More changes:
- integration tests use `non-interactive` mode
- added `--watch` parameter to the `non-interactive` mode to run a command with an interval
- interactive mode asks for confirmation on exit (to avoid accidental node termination)
- added the `watch` command to the interactive mode

Motivation and Context
---
Improve the user experience of the node's shell.
Simplify maintenance and adding new commands.
Also periodically status prnting breaks the rustyline and prompt chars not printed properly.

How Has This Been Tested?
---
Manually
