+++ applications/tari_base_node/Cargo.toml
@@ -24,8 +24,10 @@ tari_shutdown = { path = "../../infrastructure/shutdown" }
 tari_utilities = { git = "https://github.com/tari-project/tari_utilities.git", tag = "v0.3.1" }
 
 anyhow = "1.0.53"
+async-trait = "0.1.52"
 bincode = "1.3.1"
 chrono = { version = "0.4.19", default-features = false }
+clap = { version = "3.1.1", features = ["derive"] }
 config = { version = "0.9.3" }
 crossterm = "0.22"
 derive_more = "0.99.17"
@@ -34,10 +36,11 @@ futures = { version = "^0.3.16", default-features = false, features = ["alloc"] 
 log = { version = "0.4.8", features = ["std"] }
 log-mdc = "0.1.0"
 num_cpus = "1"
+nom = "7.1.0"
 regex = "1"
 rustyline = "9.0"
 rustyline-derive = "0.5"
-strum = "0.22"
+strum = { version = "0.22", features = ["derive"] }
 strum_macros = "0.22"
 thiserror = "^1.0.26"
 tokio = { version = "1.11", features = ["signal"] }
