+++ applications/tari_base_node/src/commands/mod.rs
@@ -1,7 +1,7 @@
 mod args;
 pub mod cli;
-pub mod command_handler;
+pub mod command;
+pub mod nom_parser;
 pub mod parser;
-pub mod performer;
 pub mod reader;
-mod status_line;
+pub mod status_line;
